Update cast event types and docs
